Removed action and reducer index files
