+++ src/actions/settingsActions.js
@@ -10,6 +10,32 @@ export const GET_USER_DATA_SUCCESS = 'GET_USER_DATA_SUCCESS'
 export const GET_USER_DATA_START = 'GET_USER_DATA_START'
 export const UPDATE_LOGGED_IN_USER = 'UPDATE_LOGGED_IN_USER'
 
+//Favorites
+export const SET_FAVORITES_START = "SET_FAVORITES_START";
+export const SET_FAVORITES_FAILURE = "SET_FAVORITES_FAILURE";
+export const SET_FAVORITES_SUCCESS = "SET_FAVORITES_SUCCESS";
+
+export const ADD_FAVORITE_START = "ADD_FAVORITE_START";
+export const ADD_FAVORITE_FAILURE = "ADD_FAVORITE_FAILURE";
+export const ADD_FAVORITE_SUCCESS = "ADD_FAVORITE_SUCCESS";
+
+export const REMOVE_FAVORITE_START = "REMOVE_FAVORITE_START";
+export const REMOVE_FAVORITE_FAILURE = "REMOVE_FAVORITE_FAILURE";
+export const REMOVE_FAVORITE_SUCCESS = "REMOVE_FAVORITE_SUCCESS";
+
+export const setFavorites = (favorites, userID) => dispatch => {
+
+	//TODO: Add eddpoint to set a user's favorites, have endpoint return the new list of favorites on success
+	console.log("ACTION SETTING FAVS", favorites);
+	dispatch({ type: SET_FAVORITES_START });
+	//hit endpoint POST userID and favorites
+	//then
+	dispatch({ type: SET_FAVORITES_SUCCESS, payload: favorites });//payload: res.data
+	//catch
+	//dispatch({ type: SET_FAVORITES_FAILURE, payload: error });
+  
+  }
+
 export const fetchEditAccount = (id, newInfo) => dispatch => {
 	dispatch({ type: FETCH_EDITACCOUNT_START })
 	axiosWithAuth()
